Add web-prover to the stack file of literate-unitb
o Bring in the web prover project under stack so that the build and test
commands will be performed for the web prover project as well.

o Fix the tests of the web prover project and make them pass.

TODO:

o Write a test case in the web project's test suite for testing new
well-definedness feature.