Separate rendering from proof checking / Render complete form
o Rename "Declarations" to "Constants" on the interface.

o Render all the fields of the form (theories, declarations, assumptions
, and goal).

o Separate LaTex rendering from the call to the proof checking. This way
we call LaTeX right away when the user submits the form and we don't
have to wait to hear from the prover to start rendering.

o Derive lenses for ProofForm.

o Derive Functor for ProofForm and implement an instance of
FunctorWithIndex and use `imap toStringLi` in pfStringToPfStringLi.

TODO:

o Render goal of presets and other examples.

o Handle the cases where theories, declarations or assumptions are each
or all empty.

o Implement typeToSet for RecordSort.